+++ Therapy.Domain/DTOs/WorkoutExercise/WorkoutExerciseUpdateDTO.cs
@@ -3,8 +3,8 @@ namespace Therapy.Domain.DTOs.WorkoutExercise
     public class WorkoutExerciseUpdateDTO
     {
         public int? ExerciseId { get; set; }
-        public int Sets { get; set; }
-        public int Reps { get; set; }
+        public Int16 Sets { get; set; }
+        public Int16 Reps { get; set; }
         public TimeSpan? Duration { get; set; }
     }
 }
